# web/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OTRUST Blockchain Timestamp API",
    "version": "1.0.0",
    "description": "Free, zero-knowledge REST API for blockchain-anchored timestamping. Prove data existed at a specific time using Bitcoin blockchain and OpenTimestamps.",
    "contact": {
      "name": "OTRUST",
      "url": "https://www.otrust.eu",
      "email": "hello@otrust.eu"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://www.otrust.eu"
    }
  ],
  "paths": {
...
    "/api/v1/platform/me": {
      "get": {
        "summary": "Platform identity for API key",
        "tags": ["Platform"],
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": { "description": "Org and key metadata" },
          "401": { "description": "Missing or invalid API key" }
        }
      }
    },
...
    "/api/v1/platform/organizations": {
      "post": {
        "summary": "Create organization (admin)",
        "tags": ["Platform"],
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "responses": {
          "201": { "description": "Organization created" },
          "409": { "description": "Idempotency key conflict" }
        }
      },
      "get": {
        "summary": "List organizations (admin)",
        "tags": ["Platform"],
        "parameters": [
          {"$ref": "#/components/parameters/PageLimit"},
          {"$ref": "#/components/parameters/PageCursor"}
        ],
        "responses": {
          "200": {
            "description": "Organization list with an optional next_cursor"
          }
        }
      }
    },
    "/api/v1/platform/webhooks/endpoints": {
      "post": {
        "summary": "Register webhook endpoint",
        "tags": ["Platform"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "responses": { "201": { "description": "Endpoint created" } }
      },
      "get": {
        "summary": "List webhook endpoints",
        "tags": ["Platform"],
        "security": [{ "ApiKeyAuth": [] }],
        "responses": { "200": { "description": "Endpoint list" } }
      }
    },
...
  },
  "components": {
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "description": "Unique key for safely retrying a write request for 24 hours. Reusing a key with a different payload returns 409.",
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 128,
          "pattern": "^[a-zA-Z0-9._-]+$"
        }
      },
      "PageLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "PageCursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Opaque next_cursor value returned by the previous page.",
        "schema": {"type": "string"}
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer otrust_live_... or X-OTRUST-Key header"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Short-lived issuer-bound Auth token"
      }
    },
...
  },
...
}